# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: 'Autpost API'
  description: 'API documentation for the Autpost application'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:9527/api'
    description: 'Local development server'
paths:
  /access-logs:
    get:
      tags:
        - 'Access Logs'
      summary: 'Get all access logs with pagination'
      operationId: c245b593096daf0380e2c933dce6b616
      parameters:
        -
          name: page
          in: query
          required: false
          schema:
            type: integer
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
        -
          name: source
          in: query
          required: false
          schema:
            type: string
        -
          name: path
          in: query
          required: false
          schema:
            type: string
        -
          name: ip
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: 'List of access logs with pagination'
        401:
          description: Unauthenticated
      security:
        -
          bearerAuth: []
    post:
      tags:
        - 'Access Logs'
      summary: 'Create a new access log'
      operationId: 43bb8d3ba0b24cab35db2b7ba1cbb397
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - source
                - path
                - ip
                - user_agent
              properties:
                source:
                  type: string
                path:
                  type: string
                ip:
                  type: string
                user_agent:
                  type: string
              type: object
      responses:
        201:
          description: 'Access log created successfully'
        401:
          description: Unauthenticated
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
  '/access-logs/{id}':
    get:
      tags:
        - 'Access Logs'
      summary: 'Get a single access log'
      operationId: 1af8b6eaa427cdf69bd10dd5acfec4a3
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Access log details'
        401:
          description: Unauthenticated
        404:
          description: 'Access log not found'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - 'Access Logs'
      summary: 'Update an access log'
      operationId: 48e7cdebd4ad20fd87d84332c067880f
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                source:
                  type: string
                path:
                  type: string
                ip:
                  type: string
                user_agent:
                  type: string
              type: object
      responses:
        200:
          description: 'Access log updated successfully'
        401:
          description: Unauthenticated
        404:
          description: 'Access log not found'
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - 'Access Logs'
      summary: 'Delete an access log'
      operationId: 46e2c7db2b9693b64ba70100bd016437
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Access log deleted successfully'
        401:
          description: Unauthenticated
        404:
          description: 'Access log not found'
      security:
        -
          bearerAuth: []
  /login:
    post:
      tags:
        - Auth
      summary: 'Login and create a new token'
      operationId: 67d6d0fdd54fba132e1b6f55ca254183
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              type: object
      responses:
        200:
          description: 'Login successful'
        422:
          description: 'Validation error'
  /logout:
    post:
      tags:
        - Auth
      summary: 'Logout and revoke the current token'
      operationId: 28d48c13fca984400ac8de99f54ee13a
      responses:
        200:
          description: 'Logged out successfully'
        401:
          description: Unauthenticated
      security:
        -
          bearerAuth: []
  /user:
    get:
      tags:
        - Auth
      summary: 'Get the authenticated user'
      operationId: 2baab1eeb5fd49516237a38da22e7a07
      responses:
        200:
          description: 'User data'
        401:
          description: Unauthenticated
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Enter your Sanctum token'
      bearerFormat: token
      scheme: bearer
...